requirements.txt
exif==1.6.0
ExifRead==3.0.0
pypdf==3.3.0


Pillow~=9.4.0